table2sbml works with network.py
